backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0

